// src/angular/src/app/enums/enum-target.enum.ts
export enum EnumTarget {
    None = "",
    FemaleChild = "female child",
    MaleChild = "male child",
    Child = "child",
    FemaleTeen = "female teen",
    MaleTeen = "male teen",
    Teen = "teen",
    FemaleYoungAdult = "female young adult",
    MaleYoungAdult = "male young adult",
    YoungAdult = "young adult",
    FemaleAdult = "female adult",
    MaleAdult = "male adult",
    Adult = "adult",
    Family = "family"
}